Add Pathname#glob
To run a glob search relative to a directory that is already available
as Pathname instance it was previously necessary to do this:

  dir = Pathname('/some/path')n
  Pathname.glob(dir.join('*.txt'))

now it can be:

  dir = Pathname('/some/path')
  dir.glob('*.txt')

A common use case would be to do something relative to the project root
directory, which for example in Rails projects is available as
`Rails.root`. So e.g. to find all ruby files of the current Rails
project:

  Rails.root.glob('**/*.rb')